Add end of supply notification handler
