+++ source/business-workflow/source/Energinet.DataHub.MarketRoles.Infrastructure/EDIMessaging/ENTSOE/CIM/ChangeOfSupplier/EndOfSupplyNotification/EndOfSupplyNotifier.cs
@@ -12,25 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using Energinet.DataHub.MarketRoles.Application.Common.Commands;
+using System.Threading.Tasks;
+using Energinet.DataHub.MarketRoles.Application.ChangeOfSupplier.Processing.Commands.EndOfSupplyNotification;
 using Energinet.DataHub.MarketRoles.Domain.MeteringPoints;
 
-namespace Energinet.DataHub.MarketRoles.Application.ChangeOfSupplier.Processing.Commands
+namespace Energinet.DataHub.MarketRoles.Infrastructure.EDIMessaging.ENTSOE.CIM.ChangeOfSupplier.EndOfSupplyNotification
 {
-    public class NotifyCurrentSupplier : InternalCommand
+    public class EndOfSupplyNotifier : IEndOfSupplyNotifier
     {
-        public NotifyCurrentSupplier(Guid accountingPointId, Guid businessProcessId, string transaction)
+        public Task NotifyAsync(AccountingPointId accountingPointId)
         {
-            AccountingPointId = accountingPointId;
-            BusinessProcessId = businessProcessId;
-            Transaction = transaction;
+            //TODO: Add logic for generating actor notification message
+            return Task.CompletedTask;
         }
-
-        public Guid AccountingPointId { get; }
-
-        public Guid BusinessProcessId { get; }
-
-        public string Transaction { get; }
     }
 }
