+++ source/business-workflow/source/Energinet.DataHub.MarketRoles.Tests/Application/ChangeOfSupplier/Process/ChangeOfSupplierProcessManagerTests.cs
@@ -16,6 +16,7 @@ using System;
 using System.Linq;
 using Energinet.DataHub.MarketRoles.Application.ChangeOfSupplier.Processing;
 using Energinet.DataHub.MarketRoles.Application.ChangeOfSupplier.Processing.Commands;
+using Energinet.DataHub.MarketRoles.Application.ChangeOfSupplier.Processing.Commands.EndOfSupplyNotification;
 using Energinet.DataHub.MarketRoles.Application.ChangeOfSupplier.Processing.Events;
 using Energinet.DataHub.MarketRoles.Application.Common.Processing;
 using Energinet.DataHub.MarketRoles.Domain.MeteringPoints;
